+++ pyabc/__init__.py
@@ -41,7 +41,9 @@ from .epsilon import (
     MedianEpsilon,
     ListEpsilon)
 from .smc import ABCSMC
-from .storage import History
+from .storage import (
+    History,
+    create_sqlite_db_id)
 from .acceptor import (
     Acceptor,
     SimpleFunctionAcceptor,
@@ -124,6 +126,7 @@ __all__ = [
     "IntegratedModel",
     # history
     "History",
+    "create_sqlite_db_id",
     # visualization
     "visualization",
 ]