+++ pyabc/storage/history.py
@@ -8,6 +8,7 @@ import scipy as sp
 from sqlalchemy import func
 from sqlalchemy.orm import subqueryload
 from functools import wraps
+import tempfile
 import logging
 
 from .db_model import (ABCSMC, Population, Model, Particle,
@@ -59,6 +60,28 @@ def git_hash():
             git.exc.InvalidGitRepositoryError) as e:
         git_hash = str(e)
     return git_hash
+
+
+def create_sqlite_db_id(
+        dir_: str = None,
+        file_: str = "pyabc_test.db"):
+    """
+    Convenience function to create an sqlite database identifier which
+    can be understood by sqlalchemy.
+
+    Parameters
+    ----------
+    dir_:
+        The base folder name. Optional, defaults to the system's
+        temporary directory, i.e. "/tmp/" on Linux. While this makes
+        sense for testing purposes, for productive use a non-temporary
+        location should be used.
+    file_:
+        The database file name. Optional, defaults to "pyabc_test.db".
+    """
+    if dir_ is None:
+        dir_ = tempfile.gettempdir()
+    return "sqlite:///" + os.path.join(dir_, file_)
 
 
 class History:
@@ -214,7 +237,7 @@ class History:
             -> (pd.DataFrame, np.ndarray):
         """
         Returns the weighted population sample for model m and timepoint t
-        as a pd.DataFrame.
+        as a pandas DataFrame.
 
         Parameters
         ----------
@@ -263,13 +286,11 @@ class History:
 
         Parameters
         ----------
-
         t: int, optional (default = -1)
             Population index.
 
-        Returnes
-        --------
-
+        Returns
+        -------
         model_names: List[str]
             List of model names.
         """
@@ -812,7 +833,7 @@ class History:
 
         nr_particles_per_population: pd.DataFrame
             A pandas DataFrame containing the number
-            of particles for each population
+            of particles for each population.
 
         """
         query = (self._session.query(Population.t)